NO-JIRA : [Java Config] Added connection last io time, vhost get exchange types... and made TODOs more explicit
git-svn-id: https://svn.apache.org/repos/asf/qpid/branches/java-config-and-management@1302127 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/java/broker/src/main/java/org/apache/qpid/server/management/ServerConnectionMBean.java
@@ -84,7 +84,7 @@ public class ServerConnectionMBean extends AbstractAMQManagedConnectionObject
 
     public Date getLastIoTime()
     {
-        return new Date(_serverConnection.getLastIoTime().longValue());
+        return new Date(_serverConnection.getLastIoTime());
     }
 
     public Long getMaximumNumberOfChannels()
